wl#7332 follow-up. Assert fixed.
A duplicate state message was detected, like
  09:42:05.843143Z 0 [ERROR]  'Unexpected State message from (16777343,16704)'
which appeared to be result of insufficient cleanup from that member former
session with the cluster.
At resuming (with START GCS_REPLICATION) the member was believed in that it's
not a joiner.

Fixed with it's former presence in the cluster is wiped off at Protocol::open_session()
along with the rest of cleanup.

Few asserts and more comments are added. max_view_id_p_id computation
is slightly refined to keep the id of not the first of max-view-id
senders but the last. There's no implication on correctness, and it
done merely to help debug comparison in a following loop inside
process_state_messages().