Bug#17547557: MANY TIME_WAIT SOCKETS DURING NDBMTD RECOVERY
A socket enters TIME_WAIT state for 4 minutes by default after it is 
closed by the user. Given a TCP link setup between two hosts, the link 
close may be initiated by either (or both) of the machines. The 
TIME_WAIT state is intended to give time for reordered/delayed packets 
in the network to arrive (or exceed their TTL) before the socket can 
be reused.

1) Each node (Api, MGMD + Ndbd) in the cluster has a TransporterRegistry
2) Each TransporterRegistry is responsible for establishing and managing 
cluster connections
3) Data nodes act as TCP Servers to Api nodes and to each other, and 
clients to each other
7) The start clients thread periodically attempts to connect to each 
node which it is not currently connected to (Up to 10 times a second)
8) While a data node is starting, it accepts TCP connections from Api 
nodes, 'authenticates' them, and checks their node ids, but then 
*immediately disconnects them*.
9) The start_clients_thread does not differentiate between successful 
TCP connections which were disconnected and unsuccessful TCP 
connections, and retries in the next cycle.

The effect of this is that each Api node (or connection in a connection-
pooled setup) can connect and be disconnected up to 10 times a second 
from each ndbmtd. The ndbmtd can put up to 10 sockets/s per Api client 
into TIME_WAIT state.

This fix modifies the protocol and code to avoid constant disconnect + 
reconnect during node restart. The server-side of connection would 
instead of disconnecting the client-side, send back a WAIT message.That 
message would then also have to be resent every N seconds (so that the 
connection isn't declared dead).