wl#7332 follow-up. Assert fixed.
An assert fired, see

   http://fimafeng11.no.oracle.com:8080/view/NCARVALH/job/mysql-trunk-wl6819/29/console

to designate a failure to State message exchange protocol.
A duplicate state message was detected,
    09:42:05.843143Z 0 [ERROR]  'Unexpected State message from (16777343,16704)'
which could be caused by two factors:
a. by the result of insufficient cleanup from that member former
   session with the cluster so that
   at resuming (with START GCS_REPLICATION) the member was believed in that it's
   not a joiner, and
b. by (natural) unawareness of Joiner member of what it precisely the number of State
   Messages to wait.

Fixed with two measures taken:

a. it's former presence in the cluster is wiped off at Protocol::open_session()
   along with the rest of cleanup, and
b. the Joiner made to ignore the superfluous messages.
   Such messages are exact copies of those that already arrived.

One warning's details are corrected (process id of the local joiner).
Few asserts and more comments are added. max_view_id_p_id computation
is slightly refined to keep the id of not the first of max-view-id
senders but the last. There's no implication on correctness, and it
done merely to help debug comparison in a following loop inside
process_state_messages().