bug#18496153 	EXCESSIVE PHASE 101 DURATION (OR STUCK IN PHASE 101)
Patch adding configurable time limit on duration of phase 101 wait for
subscribed nodes to connect.

Duration defaults to 2 minutes, but can be configured in milliseconds 
using the RestartSubscriberConnectTimeout parameter.

Duration of 0 is unlimited.

During wait, details of the missing subscribers will be logged every
10 seconds.

When wait expires, missing subscribers will be disconnected from all
data nodes, which should allow the restart to proceed.

If this fails to clear the problem, a second wait duration will occur,
with detail logs every 10s.

If this second duration expired, the node will shutdown.