WL#6819: GCS Replication: Overview
Post fix for lock and test problems.
      
A lock was being copied instead of being accessed by reference under
the replication threads API for GCS. This could cause the lock to
block as it could be locked when copied and never notified of states 
changes.
A small redundant value attribution is also removed for clarity.

In terms of testing, a problem existed in the wait for node status
include file where the node_id  was considered null when it started
by 0.