WL#6819: GCS Replication: Overview Post fix: relay log purge
Due to bad coding on MYSQL_BIN_LOG::reset_logs, the GTID set that is
cleared when this method is invoked is one on the active_mi relay log
and not the one on the calling relay log info object.
For now this problem is mitigated by reallocating this instruction to
the Relay_log_info::purge_relay_logs method.