WL#6819: GCS Replication: Overview Post fix: how to problems fix
When doing a hands on on GCS some problem were found when stopping the
plugin on error cases resulting on server crashes.  This was due to
cleaning operations on the donor connection that wasn't yet
initialized. This is now avoided.
Problems where also found when doing repeated recovery rounds on one
node after executing RESET MASTER on that same node.  When doing a
RESET MASTER and cleaning the data on the node, pre connection relay
log purge is needed on the recovery donor connection.