WL#6842 Transaction Applier Module
(Reset Master support post fix)

Due to previous changes, the executed relay log purging executed in
the plugin no longer guaranteed that the received set variable was
successfully cleared.
By dividing different purge scenarios and executing a double purge log
on initialization situations the problem is solved (in the slave this
is not needed due to the rotate received by the master).