BUG#20261437: CRASH IN SAFE_MUTEX_LOCK WITH GROUP REPLICATION
Group Replication requires that all tables have a primary key.
WL#7988 was implemented to ensure that only valid tables were used,
although one of conditions was not checking if primary key index was
more or equal to MAX_KEY, which was causing some invalid tables not
to be detected.

Now that condition, that primary keys must be less than MAX_KEY is
also checked, and invalid tables detected before execution.