BUG#20268486: ASSERTION INT BINLOG_CACHE_DATA::FINALIZE(THD*, LOG_EVENT*) BUG#20266984: WL#8242: ERROR - FAILED TO REGISTER FOR CERTIFICATION OUTCOME
When Group Replication plugin was unable to broadcast a transaction
write set and data not all internal plugin state was properly
cleaned. This invalid internal state was causing errors on the next
transaction broadcast, making it to fail.

The issue is now fixed by cleaning transaction certification ticket
when errors do happen.