Bug #20204537: WL#8242: RECOVERY STALLS DUE TO TRANSACTION EXECUTED ON RECOVERING NODE
During recovery if transactions are executed by a user, they can
acquire locks and then block waiting for certification that will only
come when recovery ends. The problem with this is that this locks can
lead to the server hanging as recovery cannot advance.

To solve this, during recovery user executed transactions are now
aborted.
A case is also added for error cases where the plugin goes offline,
being transactions also aborted in this case.