Bug#20538279 MYSQL IGNORING REPLICATION USERS PASSWORD OPTION AFTER RESTART
When the server starts, status vars acquire their values from the
option file without invoking the associated check and update
functions. The recovery password relies on these functions to set
its value without turning the it visible to the user relying on a
dummy internal var that never changes its value.

As a solution, when the server starts and the plugin is being
initialized, we use the dummy password to fetch the value that should
not be there, cleaning its value afterwards.