HOPSWORKS-2313
Changed defaults for
ClassicFragmentation: true -> false
SpinMethod: StaticSpinning -> LatencyOptimisedSpinning
MinWriteSpeed: 10M -> 2M
MaxDiskWriteSpeed: 20M -> 4M
AutomaticThreadConfig: false -> true

The ClassicFragmentation set to true and AutomaticThreadConfig
set to false is for backwards compatability of NDB Cluster.
iRoNDB focus on new users and thus use the most modern setup
as the default configuration.

We set the default for spinning to LatencyOptimisedSpinning since
iRoNDB is focused on Cloud installations. This will give the best
latency when possible with very little extra use of power.

We decrease the setting of MinDiskWriteSpeed and MaxDiskWriteSpeed
since EnablePartialLcp and EnableRedoControl will both ensure that
we will always write enough LCP, so we decrease the minimum disk
write speed to decrease the overhead in low usage scenarios.