HOPSWORKS-2313
Introduced new configuration variables:
AutomaticMemoryConfig
TotalMemoryConfig
AutomatedMemoryConfig is set by default
By default memory size is retrieved from the OS.
Setting TotalMemoryConfig overrides this. AutomatedMemoryConfig
requires a minimum of 8GB of memory available.

When this is set we change the defaults of SchemaMemory to the following:
MaxNoOfAttributes=500000
MaxNoOfTriggers=200000
MaxNoOfTables=8000
MaxNoOfOrderedIndexes=10000
MaxNoOfUniqueHashIndexes=2300

In this mode the default action is to get the memory from the OS
using the NdbHW module. One can also set it using a new config
variable TotalMemoryConfig. This is mainly intended for testing.

The default will work very well when starting an NDB data node in
a cloud setting. In this case the data node will use the entire
VMs memory and CPU resources since also AutomaticThreadConfig is
default.

Thus using this setting it is not necessary to set anything since
even DataMemory and DiskPageBufferMemory is set based on the total
memory available.

The way it works is that the following parameters now have a 0
default. 0 means that it is automatically calculated.

RedoBuffer
UndoBuffer (this is a new config parameter introduced to make it
            possible to change the Undo buffer size).
LongMessageBuffer

RedoBuffer will get 32 MByte per LDM
UndoBuffer will get 32 MByte per LDM
LongMessageBuffer will get 12 MByte per thread used, but the first
  thread will get 32 MByte.
TransactionMemory will get 300 MByte + 60 MByte per thread
SharedGlobalMemory will get 700 MByte + 60 MByte per thread

We avoid using memory that is required by the OS. We avoid using
1800 MByte plus 100 MByte for each block thread.

SchemaMemory is around 1 GB in a small cluster and increases
with increasing cluster size and increasing node sizes.

We will compute the use of:
Restore Memory
Schema Memory
Packed Signal Memory
OS memory usage
Send Buffer Memory
Transaction Memory
Shared Global Memory
NDBFS Memory
Static memory (in various objects)
Redo Buffer
Undo Buffer
Backup Page Memory
Job Buffers

Also Page_entry objects used by DiskPageBufferMemory is taken into
account.

After computing these buffers, schema memory and various internal
memory structures and the memory we want to avoid using such that
the OS also have access to some memory, the remaining memory is
for DataMemory and DiskPageBufferMemory. 90% of the remaining memory
is given to the DataMemory and 10% to the DiskPageBufferMemory.

In addition we have changed the default of MinFreePct to 10%.
This provides some possibility to increase sizes of buffers if
necessary, this can also be decreased down to 5 and possibly
even down to 3-4 to enable more memory for database rows.

One can also set DiskPageBufferMemory to increase or decrease
its use. This will have a direct impact on the memory available
for DataMemory.

Removed a serious bug in Dblqh where DEBUG_FRAGMENT_LOCK was set.
This increased the size of each fragment record by 4kB.

Increased the hash size for foreign keys in Dbtc to 4096 from 16.