HOPSWORKS-2313
Ensured that UndoBuffer configuration parameter is used if defined or
if using automatic memory configuration when defining the UNDO log.
The memory is added to TransactionMemory and thus can be used for
this purpose if no Undo log is created.

Fixed a problem with calculating hw_memory_size on Linux.
Fine tuned some memory parameters, added a bit more constant overhead
to OS, but decreased overhead per thread.

Fixed a problem with running with 1 thread for LongMessageBuffer.

Fixed some review comments.

Fixed compute_os_memory based on OS overhead in a cloud VM.