Bug#32354817 CRASH IN HA_NDBCLUSTER HANDLER INTERFACE WHEN PREPARING A PUSHED JOIN
The pushed join handler interface maintain a set of 'join-nest', implemented as
a bitmap of the table'ids being a member of the inner- and outer- join-nests.

When we have completed the handling of the last inner-table in a join-nest
we return to the outer-nest, which then becomes the inner-nest for the
next table(s).

The existing logic for reconstructing the inner_nest we returned to, relied
upon that that we could use the m_tables[].m_upper_nest set to recalculate
the inner_nest map we returned to. That logic had an implicit requirement that
the m_upper_nest bitmap was 'const' while we handled the tables in the inner_nest
we are about to exit from.

However, in some (rare) cases we will recalculate the m_upper_nest bitmap
in ndb_pushed_builder_ctx::is_outer_nests_referable(), thus breaking the
above assumption.

This patch change the logic for recalculating the inner_nest:
- Find the last previous table in the nest we return to (the new inner_nest)
- The inner_nest will then be the m_inner_nest of this table + its table id

Note ^^ that the m_inner_nest maps are defined such that they only contain
the previous tables being member of the same inner_nest as this table.

Reviewed By: Arnab Ray <arnab.r.ray@oracle.com>