HOPSWORKS-2353
In the code to setup automatic thread config the code works to
find the correct CPUs to bind to. However missing is the setting
of m_bind_type which is what the init_thread code looks for to
discover whether to actually perform the binding.

Needed this to be added to all thread types.