HOPSWORKS-2354
When a log queue entry is to be executed, it uses the
method writePrepareLog. This method does however simply
put it backin the log queue. This method is not designed
to write from the REDO log queue.

This can be simply fixed by understanding that one arrives
in this method with log part locked and that the
logPartState is Active.