HOPSWORKS-2364
Currently there is an adaptive algorithm to decide how many objects to keep
in cache.This algorithm will keep too many objects if there are occasional
usage where a lot ofoperations is part of a transaction.

The solution is to never keep more than 128 objects in each free object list
in theNdb object.