Bug #31925977 NDB : REDO_ALERT_STATE CALCULATION INCORRECT WITH > 1 LOG PART PER LDM
The Redo Alert State mechanism added in 7.6 as part of the Partial LCP
provides an indication of the redo fill level.

Keeping redo space available avoids the need to reject transactions that
modify data.  The redo state mechanism is used to determine whether the
Local Checkpoint (LCP) should be sped up, so that redo space can be
reclaimed sooner.

When gathering redo state information, the code was gathering information
for the most filled redo log part, but also gathering the volume of
data in bytes written to *all* redo log parts managed by an LDM.

This resulted in overestimation of the risk of redo exhaustion, and high
redo alert states being set incorrectly, affecting disk write speed and
LCP resource consumption.

The fix is to gather the bytes written only to the part with the most
bytes written in the period, keeping the dimensionality of the
calculations to 1 redo log part in cases where LDMs have more than 1
part.

Approved by : Ole John Aske <ole.john.aske@oracle.com>