HOPSWORKS-2394
When executing a queued REDO log we call get_table_frag_record.
This call is used to get the instance number and actually can fail
if the fragment record doesn't belong to the owner of the REDO
log. So the ndbrequire is not correct. Actually better to separate
this call into a new function get_table_frag_instance that avoids
getting table and fragment pointers.