HOPSWORKS-2400
Currently the size of the UNDO log buffer is limited to 600 MB.To set up RonDB on a Metal Server with 96 CPUs requiresaround twice the amount using the automatic sizing of buffersand memories.

The problem is a small map in LGMAN, by increasing this from40 to 1200 we can support UNDO log buffers of up to 16 GBytewhich is sufficient to even handle a data node with 1024 CPUs.