HOPSWORKS-2394
Debugging of Queueing of log records

When executing queued log records with 3 LDMs the wrong path
in the code was choosen which resulted in the signal being
executed by the wrong block instance which led to problems.

In addition a check that the logPartState was ACTIVE was not
detailed enough, required a bit more details since we set the
state to IDLE before executing the REDO log record and not
after to ensure that we don't need to keep the mutex for too
long time.

Changed the checks to use m_use_mutex_for_log_parts that is
properly calculated to discover if we can have the operations
for a log record in a different LDM instance.