HOPSWORKS-2407
ndb_import can hang when there are NDB errors in loading a large
CSV file. In this case the execution threads stops progressing
due to the failures. However the CSV input threads and the output
workers are all stuck in trying to send rows to the next thread
in the pipeline (output threads for CSV input threads and execution
threads for output threads). They are stuck in this state since
there are no checks for stop processing in this code path.

Added checks for stop processing in those paths, but also changed
such that all errors will lead to stopping the entire ndb_import
program and not just the current job being executed. It is better
to have a program with a consistent error behaviour rather than
trying to continue processing after a failure.

Changed some parameters for ndb_import to make it more responsive
and quicker to change jobs. Some of those led to dramatic decrease
of time to perform import jobs.