HOPSWORKS-2418
ClusterJ lacked support for Primary keys using Date columns.
Removed this limitation by removing check for that this is not
supported. In addition required the addition of a new object
to handle Date as Primary Key based on the handling of Date
columns in non-key columns.

Required also a new test case in the clusterj testsuite.
This required fairly significant rework since the current testsuite
was very much focused on testing all sorts of column types, but
not very much focused on testing of different data types in primary
key columns. This meant that most of the methods had to be changed
a bit and the whole idea with reuse of code didn't work so well.

A similar approach is now also required for a set of other data
types such as YEAR, DATETIME and so forth.