HOPSWORKS-2426
To support minimal configurations down to 2GB memory space we made the
following changes:
1) Make it possible to set TotalMemoryConfig down to 2GB
2) Remove OS overhead when setting TotalMemoryConfig
3) Decrease the SendBufferMemory from 8M to 4M

Using the following settings:
NumCPUs=4
TotalMemoryConfig=3G
MaxNoOfTables=2000
MaxNoOfOrderedIndexes=1000
MaxNoOfUniqueHashIndexes=1000
MaxNoOfAttributes=20000
MaxNoOfTriggers=30000
TransactionMemory=200M
SharedGlobalMemory=400M

we will be able to fit nicely into 3GB memory space and out of this
a bit more than half is used by DataMemory and DiskPageBufferMemory.

AutomaticMemoryConfig computation will be affected by the settings
of MaxNoOfTables, MaxNoOfOrderedIndexes, MaxNoOfUniqueHashIndexes,
MaxNoOfAttributes, MaxNoOfTriggers and also the setting of
TransactionMemory and SharedGlobalMemory.

NumCPUs is used to ensure that we use a small configuration when
it comes to the number of threads if the test is started on a large
machine with lots of data nodes (can be useful for testing).