HOPSWORKS-2431
We have set MinDiskWriteSpeed and MaxDiskWriteSpeed very low in the default
configuration. These values are intended as a minimum disk write speed for
LCP processing even in the context of a lower requirement from the adaptive
algorithm deciding the disk write speed.

However there is also an algorithm to decrease the LCP speed to its minimum.
To avoid that this minimum is not below what is reasonable to process even in
the context of CPU overload or disk overload we hardcode the parameters for
disk write speed minimum, but do this per LDM thread rather than the total.

The min disk write speed is set to 2 MByte, the max to 3 MByte, the max during
another nodes restart is set to 12 MByte and the max during our own restart is
set to 50 MByte. All numbers per LDM thread.