HOPSWORKS-2441
Decouple Send thread mutex and Send thread buffers mutex
--------------------------------------------------------
One mutex is used to protect the list of free send buffers. This mutex is acquired
when a block thread is acquiring send buffers. Each block thread is connected to
a send thread. It acquires send buffers from the pool of send buffers connected to
this send thread. Whenever sending occurs the send buffer is returned to the pool
owned by the send thread. This means that a block thread will always acquire
the buffers from its send thread buffers. But it will return job buffers to the pool
connected to the transporter the message is sent to.

This means that imbalances can occur, this imbalance is fixed when seizing from
the global list, if no send buffers are available we will try to "steal" from neighbour
pools.

The actual pool of send buffers are protected by a mutex. This is not the same
mutex that is used to protect the send thread. This send thread mutex is used
to protect the list of transporters waiting to be served with sending.

However in the current implementation we hold the send thread mutex when
releasing to the send buffer pool. This means that we hold two hot mutexes
at the same time. This clearly is not good for scaling RonDB to large instances.
To avoid this we decouple the release_global call in assist_send_thread as
well as the release_chunk call in run_send_thread from the send thread mutex.

The release_global call can simply wait to the end of the assist_send_thread
call. We will normally not spend so much time in the assist_send_thread call.
The call to release_chunk in the send thread can be integrated in handle_send_trp
to avoid that we need to release and seize the send thread mutex an extra time
in the send thread loop.