HOPSWORKS-2442, HOPSWORKS-2443
Block threads are very eager to assist with sending. In situation where we have
large data nodes with more than 32 CPUs this will lead to a very busy sending.
This leads to sending chunks of data that is too small. This leads both to contention
on send mutexes and it leads to sending small messages that hurts the receivers
and makes them less efficient.

To avoid this a more balanced approach on sending is required where block
threads assist sending, but as load goes up the assistance becomes less urgent.

In this patch we use the implementation of MaxSendDelay and turn it into an
adaptive algorithm. To create this adaptive algorithm we add another CPU
load level called HIGH_LOAD. This means that we have LIGHT_LOAD up to 30%
load, MEDIUM load is between 30% and 60%, HIGH LOAD is between 60% and
75% and above that we have OVERLOAD_LOAD.

We implement two new mechanisms, the first is max_send_delay. This is
activated when threads starts reaching HIGH_LOAD. This means that when
a block thread puts a transporter into the send queue, it will have to
wait a number of microseconds before the send is allowed to occur. This
gives the send a chance to gather other sends before the send is
actually performed. The max send delay is larger for higher loads and
for larger data nodes. It is never set higher than 200 microseconds.

The second mechanism is the minimum send delay. Each time we send and the
min send delay is nonzero, we set the transporter in a mode where it has
to wait to send even if a send is scheduled. This setting normally isn't
affecting things when max_send_delay is set. So it acts to secure against
overloading the receivers when our load is still not at HIGH_LOAD.
This is at most set to 40 microseconds.

In the past up to 4 block threads could be choosen to assist the send
threads. This mechanism had a bug in resetting the activation. This is
now fixed. In addition the new mechanisms is only providing for main
and rep threads to assist the send threads. In addition the main and
rep threads only do when the send threads are at a high load. Once
we have entered this mode, we will wait to remove the assistance using
a hysteresis. We have two levels of send thread support, the first adds
the main thread and the second level adds also the rep thread. Neither
of those will assist when they are reaching up to MEDIUM_LOAD level.
This should be fairly uncommon for those threads except in some special
scenarios.

By default main and rep threads are now not assisting send threads
at all.

Block threads will only send to at most one transporter when reaching
MEDIUM_LOAD, as before they will not assist send threads at all after
reaching the OVERLOAD_LOAD level.

The main and rep threads are activated using the WAKEUP_THREAD_ORD
signal, when this is sent the main/rep thread is setting nosend=0
and once every 50ms it is restored to nosend=1. Thus it is necessary
to constantly send WAKEUP_THREAD_ORD signals to those threads to
keep them waking up to assist send threads.

To handle this swapping between nosend 0 and 1, we introduce a
m_nosend_tmp that is the one used to decide whether to perform
assistance or not, the configured value is in the m_nosend
variable.

To decide whether send thread assistance is required we modify
the method to calculate send thread load such that it uses
a parameter with number of milliseconds back we want to
track. We use 200 milliseconds to decide on send thread
assistance.

The configuration parameter MaxSendDelay is deprecated and no longer
used. All of its functionality is now automated.

As a first step to remove the recover thread from disturbing the
block threads we make sure that it never performs any send thread
assistance. This assistance will disturb things, the recover threads
still wake up and use about 0.1-0.2% of a CPU, this should also be
removed eventually or handled in some other manner.

Also ensured that recover threads are not performing any spinning.

Handling send thread assistance until done is now only handled by
main and rep threads. Thus we changed the setting of the pending_send
variable.

Removed incrementing the m_overload_counter when calling
set_max_delay. This should only be incremented when calling
the set_overload_delay function.

Added some comments on this new functionality.

Fixed some problems with lagging timers. Important to perform
do_send also when this happens.