HOPSWORKS-2365
Support ACTIVATE NODE and DEACTIVATE NODE and change of hostname
in a running cluster.

Previously setting NoOfReplicas was something that cannot change.
This feature doesn't change that. Instead it makes it possible to
set up your cluster such that it can grow to more replicas
without reconfiguring your cluster.

The idea is that you set NoOfReplicas to 3 already in the initial
config.ini. Even 4 is ok if desirable. This means that 3 Data nodes
need to be configured. However only 1 of them is required to be
properly defined. The other data nodes simply set the node id
and set NodeActive=0. This indicates that the node is deactivated.
Thus the node isn't possible to include into the cluster. All nodes
in the cluster are aware that this node exists, but they do not
allow it to connect since the node is deactivated.

This has the advantage that we don't expect a node to start when
we start another data node, we don't expect it to start when
waiting in the NDB API until the cluster is started.

This has the added advantage that we can temporarily increase
the replication from e.g. 2 to 3. This could be valuable if
we want to retain replication even in the presence of restarts,
software changes or other reasons to decrease the replication
level.

A node can be activated and deactivated by a command in the MGM
client. It is a simple command:

4 activate

This command will activate node 4.

Here is the output from show for a cluster with 3 replicas, 2
of them active, it has the possibility to run with 2 MGM servers,
but only one is active, it has the possibility to run with 4
API nodes, but only one of them is active:

ndb_mgm> show
Cluster Configuration
---------------------
[ndbd(NDB)]     3 node(s)
id=3    @127.0.0.1  (RonDB-21.04.0, Nodegroup: 0, *)
id=4    @127.0.0.1  (RonDB-21.04.0, Nodegroup: 0)
id=5 (not connected, node is deactivated)

[ndb_mgmd(MGM)] 2 node(s)
id=1    @127.0.0.1  (RonDB-21.04.0)
id=2 (not connected, node is deactivated)

[mysqld(API)]   4 node(s)
id=6 (not connected, node is deactivated)
id=7 (not connected, node is deactivated)
id=8 (not connected, node is deactivated)
id=9 (not connected, accepting connect from any host)

Another important feature that is brought here is the
possibility to change the hostname of a node. This
makes it possible to move a node from one VM to
another VM.

For API nodes this also makes it possible to make
RonDB a bit safer. It makes it possible to configure
RonDB with all API nodes only able to connect from a
specific hostname. It is possible to have many more
API nodes configured, but those can not be used until
someone activates them. Before activating them one
should then set the hostname of those nodes such that
only the VM that is selected can connect to the cluster.

Thus with this feature it becomes possible to start with a
cluster running all nodes on one VM (ndb_mgmd, ndbmtd and mysqld).
Later as the need arises to have HA one adds more VMs and more
replicas. Growing to more node groups as an online operation is
already supported by RonDB. With this feature it is no longer
required that all node groups at all times have the same level of
replication. This is possible by having some nodes deactivated
in certain node groups.

The benefit of different replication within different node groups
can be used when reorganising the cluster or when one node is
completely dead for a while, or we simply want to spend less
money temporarily on the cluster.

When moving a node to a new VM there are two possibilities for this.
One possibility is that we use cloud storage as disks, this means
that the new VM can reuse the file system from the old VM, in
this case a normal restart can be performed. Thus the node is first
deactivated (includes stopping it), the new hostname is set for the
new VM, the node is activated and finally the node is started again
in a new VM. The other option is where the node is moved to a VM
without access to the file system of the old VM, in this case the
node need to be started using --initial. This ensures that the
node gets the current version of its data from other live nodes in
the RonDB cluster. This works for both RonDB data nodes as well
as for RonDB management nodes.

It is necessary to start a node that has been recently activated
using the --initial flag. If not used we could get into a situation
where the nodes don't recognize that they are part of the same
cluster. This is particularly true for MGM server where it is a
requirement that all MGM servers are alive when a configuration
change is performed. With this change the requirement is for all
active nodes to be alive for a configuration change to be performed.
In this work it is assumed that 2 MGM servers is the maximum number
of MGM servers in a cluster.

A new test program directory for this feature is created in the directory
mysql-test/activate_ndb.

Running the script test_activate.sh in this directory will execute the
test case for this test run. To perform this test it is necessary to
set the PATH such that it points to the binary directory where the
build placed the binaries. So e.g. if we created a build directory
called debug_build and the git tree is placed in /home/mikael/activate_2365
we would set the path in the following manner:
export PATH=/home/mikael/activate_2365/debug_build/bin:$PATH
and next we will run the
test_activate.sh script that will tell us if the test passed or not.