HOPSWORKS-2475
The code to create a CPU map to use in assigning CPU locking didn't
take into account the case with 0 LDM instances. Easiest solution was
to simply change to 1 LDM instance, the code above will get its list
of CPUs that it desired also in this case. Added a number of unit
tests to verify this feature as well.